services/analyzers/python_analyzer/requirements.txt
# Core dependencies
pika==1.3.1
grpcio>=1.71.0
grpcio-tools>=1.71.0
protobuf>=4.23.0

# AST parsing
astroid==2.15.5

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.1.0

# Testing
pytest==7.3.1
pytest-mock==3.10.0
pydantic
psycopg2-binary==2.9.9